Added Adjustable Width Tabdivs
TabDivs now adjust their width to match the parent id=wikipage div.  I
do this by writing out divs with class=vex-tabdivinit containing the
vextab mini language to the page using the Haskell plugin. I then use
index.js to get the width of the wikipage div, set the width attribute
on the class=vex-tabdivinit divs, and finally change the class of those
divs to vex-tabdiv so they get picked up by the vextab script that is
included in the script tags of the wiki page templates.